requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.7
django-environ==0.4.5
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Faker==4.1.0
mysqlclient==1.4.6
phonenumbers==8.12.5
phonenumberslite==8.12.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
